+++ packages/moonstone/Panels/Viewport.js
@@ -56,12 +56,22 @@ const ViewportBase = kind({
 		 * @type {Boolean}
 		 * @default false
 		 */
-		noAnimation: PropTypes.bool
+		noAnimation: PropTypes.bool,
+
+		/**
+		 * When `true`, the Panels will not fade-in. Instead, they will appear instantly.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
+		noFadeIn: PropTypes.bool
 	},
 
 	defaultProps: {
 		index: 0,
-		noAnimation: false
+		noAnimation: false,
+		noFadeIn: false
 	},
 
 	styles: {
@@ -75,10 +85,11 @@ const ViewportBase = kind({
 	},
 
 	computed: {
-		children: ({children, generateId}) => React.Children.map(children, (child, index) => {
+		children: ({children, generateId, noFadeIn}) => React.Children.map(children, (child, index) => {
 			return React.cloneElement(child, {
 				containerId: child.props.containerId || generateId(index, 'panel-container', Spotlight.remove),
-				'data-index': index
+				'data-index': index,
+				noFadeIn: typeof child.props.noFadeIn === 'undefined' ? noFadeIn : child.props.noFadeIn
 			});
 		}),
 		enteringProp: ({noAnimation}) => noAnimation ? null : 'hideChildren'
@@ -86,6 +97,7 @@ const ViewportBase = kind({
 
 	render: ({arranger, children, enteringProp, index, noAnimation, ...rest}) => {
 		delete rest.generateId;
+		delete rest.noFadeIn;
 
 		const count = React.Children.count(children);
 		invariant(